requirements.txt
streamlit
requests
paddlepaddle
paddleocr
pandas
python-dotenv
opencv-python-headless
